feat!: add c29 pow algorithm (#7398)

+++ applications/minotari_app_grpc/src/conversions/consensus_constants.rs
@@ -110,7 +110,7 @@ impl From<ConsensusConstants> for grpc::ConsensusConstants {
         #[allow(deprecated)]
         Self {
             coinbase_min_maturity: cc.coinbase_min_maturity(),
-            blockchain_version: cc.blockchain_version().into(),
+            blockchain_version: u16::from(cc.blockchain_version()).into(),
             future_time_limit: cc.ftl().as_u64(),
             difficulty_block_window: cc.difficulty_block_window(),
             max_block_transaction_weight: cc.max_block_transaction_weight(),
